update cache when using Query service (#879)
I was using the Query service (https://www.apollographql.com/docs/angular/basics/services.html#Basic-example) as described in the article, and I figured out how to use it in the cache update. The query service seemed to be the way to go when you follow the guides in a linear fashion, so I wanted to share how to present the query in the proxy object.