feat: strip unncessary characters from the query
See: https://stackoverflow.com/questions/417142/what-is-the-maximum-length-of-a-url-in-different-browsers

This came up in the https://github.com/graycoreio/daffodil when trying to send GraphQl queries as GET (for caching reasons with the GraphQL API of the ecommerce system Magento 2).
The deeper a query became, the more "%20" characters were added to the URL, which eventually overflowed Chrome (and various other browsers) max URL length.

This strips the useless characters out, reduces bandwidth consumption, and saves us all from some painful edge-cases.