New `provideApollo()` and `provideNamedApollo()`
Those two providers make it simpler to provide Apollo with options and
flags to the application.

Token injections, `APOLLO_OPTIONS`, `APOLLO_NAMED_OPTIONS` and
`APOLLO_FLAGS` are not needed anymore in application code. Though they
might still be used for custom providing techniques.

`ApolloModule` is deprecated because it brings no value. Especially with
the introduction of the two providing functions.